dead code elimination: only compute affine hull in extended spaces
The affine hull is computed to avoid infinite loops where
live iterations are continuously added to the same space(s),
which may otherwise happen in case of a dependence cycle.

When elements in a new space get added,
then these may not even be part of a dependence cycle,
so there is initially no need to compute an affine hull.
For those that do turn out to be part of a dependence cycle,
an affine hull will be computed in some future round.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>