dead code elimination: try and preserve more constraints in dependence cycles
When the dead code elimination procedure considers extra live iterations
in a space for which it already has live iterations (which happens
in particular when it encounters a dependence cycle), the affine hull
of the union is computed to ensure that the procedure terminates.
The result of the affine hull computations is bounded
by the original domain constraints, but some other interesting constraints
may still get lost.

Also preserve some constraints that were valid before the extra
live iterations got added and that are not invalidated
by those extra live iterations.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>